+++ src/components/music/CheckAnswerPopper.js
@@ -1,5 +1,5 @@
 import React, { Fragment } from "react"
-import withSimpleErrorBoundary from "../util/withSimpleErrorBoundary"
+import withSimpleErrorBoundary from "../../util/withSimpleErrorBoundary"
 import { Popper, Fade, Typography } from "@material-ui/core"
 import styled from "styled-components"
