+++ src/partials/index.js
@@ -59,7 +59,6 @@ import H6 from "./Headers/H6"
 import TableOfContents from "./TableOfContents"
 import Notice from "./Notice"
 import RegistrationLink from "./RegistrationLink"
-import DropDownForAnswers from "./DropDownForAnswers"
 
 const mapping = {
   test: Test,
@@ -78,7 +77,6 @@ const mapping = {
   "table-of-contents": TableOfContents,
   "registration-link": RegistrationLink,
   "music-sheet": MusicSheet,
-  "answer-selection": DropDownForAnswers,
   "music-exercise": MusicExercises,
   banner: Banner,
   notice: Notice,
